scripts/docker/python-image/python_image_requirements.txt
astroid~=3.0.2
attrs~=23.2.0
certifi~=2023.11.17
charset-normalizer~=3.3.2
contextlib2~=21.6.0
contourpy~=1.2.0
cycler~=0.12.1
dill~=0.3.7
flake8~=6.1.0
fonttools~=4.47.2
idna~=3.6
isort~=5.13.2
jsonschema~=4.20.0
jsonschema-specifications~=2023.12.1
kiwisolver~=1.4.5
matplotlib~=3.8.2
mccabe~=0.7.0
numpy~=1.26.3
packaging~=23.2
pandas~=2.1.4
Pillow~=10.1.0
platformdirs~=4.1.0
pycodestyle~=2.11.1
pyflakes~=3.1.0
pylint~=3.0.3
pyparsing~=3.1.1
python-dateutil~=2.8.2
pytz~=2023.3.post1
PyYAML~=6.0.1
referencing~=0.32.1
requests~=2.31.0
rpds-py~=0.17.1
schema~=0.7.5
setuptools~=69.0.3
six~=1.16.0
tomlkit~=0.12.3
tzdata~=2023.4
urllib3~=2.1.0
wheel~=0.42.0